Adding support for SSE-S3 in GET and PUT path.
This patch introduces the support for SSE-S3 in GET anf PUT path.
In the PUT path, we chec if BucketEncryption is enabled for the bucket,
which received PUT call. If yes, we detemine if the encryption type is
AES256 (i.e., SSE-S3), then we fetch the KEK-ID from the bucket X-attrs
and use it to encrypt the data key. Thereafter, we call the
generate-data-key. We store the KEK-ID and the wrapped data-key in the
object xattrs.
In the GET path, we simply pull out the KEK-ID from the object xattr and
decrypt the object.
NOTE: We only support Hashicorp Vault and the SSE-S3 KMS backend.

Signed-off-by: Priya Sehgal <priya.sehgal@flipkart.com>