rgw: Adding SSE-S3 support in GET and PUT paths (using Vault as KMS)
Generating the KEK based on bucket owner UID in PutBucketEncryption.

Introducing the support for SSE-S3 in GET and PUT path.
In the PUT path, we check if BucketEncryption is enabled for the bucket.
If yes, we detemine if the encryption type is AES256 (i.e., SSE-S3),
then we fetch the KEK-ID from the bucket x-attrs and use it to wrap the
data key. Thereafter, we call generate-data-key. We store the KEK-ID
and the wrapped data-key in the object x-attrs.
In the GET path, we simply pull out the KEK-ID from the object x-attr and
decrypt the object.

Signed-off-by: Priya Sehgal <priya.sehgal@flipkart.com>